Saving progress on refactoring code
