+++ internal/models/models.go
@@ -1,0 +1,9 @@
+package models
+
+type Database interface {
+	Connect(url string) error
+	TestConnection(url string) error
+	CloseConnection() error
+	ExecuteQuery(query string) (columns []string, rows [][]string, err error)
+}
+
